// src/features/editableProfileCard/model/types/editableProfileCardSchema.ts
import { Profile } from "entities/Profile";

export enum ValidateProfileError {
    INCORRECT_USER_DATE = 'INCORRECT_USER_DATE',
    INCORRECT_AGE = 'INCORRECT_AGE',
    INCORRECT_COUNTRY = 'INCORRECT_COUNTRY',
    NO_DATA = 'NO_DATA',
    SERVER_ERROR = 'SERVER_ERROR'
}

export interface ProfileSchema {
  data?: Profile;
  form?: Profile;
  isLoading: boolean;
  error?: string;
  readOnly: boolean;
  validateErrors?: ValidateProfileError[];
}